refactor: refactor register an user
