+++ src/features/authentication/confirm-user/model/use-send-one-time-password.test.ts
@@ -2,7 +2,7 @@ import router from '@app/providers/router'
 import { checkOneTimePassword, getCurrentSession } from '@shared/auth'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { getRegisterEmailFromStorage } from '../../register-user/model/register-storage'
+import { getRegisterEmailFromStorage } from '../../model/register-storage'
 import { monitorInformation, MONITORING_EVENTS } from '../monitoring/monitoring'
 import { useSendOneTimePassword, verifyOtp } from './use-send-one-time-password'
 
@@ -30,7 +30,7 @@ vi.mock('@shared/auth', () => ({
   getCurrentSession: vi.fn()
 }))
 
-vi.mock('../../register-user/model/register-storage', () => ({
+vi.mock('../../model/register-storage', () => ({
   getRegisterEmailFromStorage: vi.fn()
 }))
