+++ src/features/authentication/model/use-auth-navigation.test.ts
@@ -1,9 +1,9 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import { getIsOtpActiveFromStorage } from '../register-user/model/register-storage'
+import { getIsOtpActiveFromStorage } from './register-storage'
 import { useAuthNavigation } from './use-auth-navigation'
 
-vi.mock('../register-user/model/register-storage', () => ({
+vi.mock('./register-storage', () => ({
   getIsOtpActiveFromStorage: vi.fn()
 }))
