+++ src/features/authentication/monitoring/monitoring.ts
@@ -1,0 +1,35 @@
+import { addBreadcrumb } from '@shared/lib'
+
+/**
+ * Monitoring category used for authentication-related events.
+ *
+ * This category groups all monitoring breadcrumbs that originate
+ * from authentication flows (e.g. login, register).
+ */
+export const CATEGORY = 'authentication'
+
+/**
+ * Collection of monitoring event identifiers related to authentication.
+ *
+ * These constants are used when sending monitoring breadcrumbs to ensure
+ * consistent event naming across the application.
+ */
+export const MONITORING_EVENTS = {
+  STORAGE_OTP_READ: 'auth.otp.storage.read',
+  STORAGE_OTP_WRITE: 'auth.otp.storage.write',
+  STORAGE_REGISTER_EMAIL_READ: 'auth.register.email.storage.read',
+  STORAGE_REGISTER_EMAIL_WRITE: 'auth.register.email.storage.write'
+}
+
+/**
+ * Records an informational monitoring breadcrumb for authentication events.
+ *
+ * This function wraps {@link addBreadcrumb} and automatically assigns
+ * the authentication monitoring category and the log level `info`.
+ *
+ * @param {string} event - The monitoring event identifier.
+ * @param {object} [data] - Optional additional data associated with the event.
+ */
+export function monitorInformation(event: string, data?: object) {
+  addBreadcrumb(event, CATEGORY, 'info', data)
+}
